third_party/WebKit/LayoutTests/inspector/sass/test-mapping-good-expected.txt

Verify WI.ASTSourceMap rebaselines.

Mapped CSS: 12
Mapped SCSS: 12
No stale entries found.

Running: testRebase
Mapped CSS: 12
Mapped SCSS: 12
No stale entries found.
